Implement pushed authorization request (PAR), when starting a login.
As described by https://datatracker.ietf.org/doc/html/rfc9126.

PAR ensures that all oauth parameters are confidentail, and makes them harder to tamper with. It does so by sending the parameters in a POST request body, instead of in the query string.
On top of that, starting the flow with a PAR request also makes it easier for us to do tracing later on, since we can easily attack a `traceparent` header to the PAR request. That is not possible if the first request happens in the browser.