Refactor continuation around opening the browser.
Previously we would create a continuation in login, that would be resolved with the JWT token once the code exchange process was done.
However, that continuation actually did two separate things:
1. Open the browser and wait for a result
2. Exchange the code for a JWT, and validate it.

This can happen in two separate continuations instead. This both makes the flow easier to follow, because control returns to the login / logout functions as soon as the browser closes; and makes it easier to trace in an upcoming PR.